break admin.views into respective folders for roles, sync, and publish
