+++ tinker/__init__.py
@@ -37,14 +37,18 @@ from tinker.faculty_bio.views import faculty_bio_blueprint
 from tinker.redirects.views import redirect_blueprint
 from tinker.heading_upgrade.views import heading_upgrade
 from tinker.e_announcements.views import e_announcements_blueprint
-from tinker.admin.views import admin_blueprint
+from tinker.admin.sync.views import sync_blueprint
+from tinker.admin.publish.views import publish_blueprint
+from tinker.admin.roles.views import blink_roles_blueprint
 
 app.register_blueprint(event_blueprint, url_prefix='/event')
 app.register_blueprint(faculty_bio_blueprint, url_prefix='/faculty-bio')
 app.register_blueprint(redirect_blueprint, url_prefix='/redirect')
 app.register_blueprint(heading_upgrade, url_prefix='/heading-upgrade')
 app.register_blueprint(e_announcements_blueprint, url_prefix='/e-announcement')
-app.register_blueprint(admin_blueprint, url_prefix='/admin')
+app.register_blueprint(sync_blueprint, url_prefix='/admin/sync')
+app.register_blueprint(publish_blueprint, url_prefix='/admin/publish')
+app.register_blueprint(blink_roles_blueprint, url_prefix='/admin/blink-roles')
