+++ tinker/admin/publish/templates/publish-home.html
@@ -21,7 +21,7 @@
     <a class="close-reveal-modal">&#215;</a>
 </div>
 
-<form action="." id="publish-search-form">
+<form action="../../templates" id="publish-search-form">
   <div class="row">
 
     <div class="large-6 columns">
