+++ tinker/admin/publish/templates/publish-more-info.html
