+++ tinker/admin/publish/templates/publish-table.html
