+++ tinker/admin/sync/templates/sync.html
